requirements.txt
click==6.6
kafka-python==1.3.1
phue==0.8
toml==0.9.2
